feat(strategy): split monitor file
