+++ pisa-proxy/proxy/strategy/src/monitors/mod.rs
@@ -1,20 +1,18 @@
 // Copyright 2022 SphereEx Authors
-//
+// 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-//
+// 
 //     http://www.apache.org/licenses/LICENSE-2.0
-//
+// 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod config;
-pub mod route;
-use route::*;
-pub mod discovery;
-pub mod readwritesplitting;
-pub mod monitors;
+pub mod connect_monitor;
+pub mod ping_monitor;
+pub mod replication_lag_monitor;
+pub mod read_only_monitor;
