+++ pisa-proxy/proxy/strategy/src/monitors/ping_monitor.rs
@@ -1,0 +1,291 @@
+// Copyright 2022 SphereEx Authors
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::collections::HashMap;
+use futures::StreamExt;
+use tokio::time::{self, Duration};
+
+use pisa_error::error::{Error, ErrorKind};
+use mysql_protocol::{client::conn::ClientConn, util::*};
+use crate::{config::MasterHighAvailability, readwritesplitting::ReadWriteEndpoint};
+use mysql_protocol::row::RowData;
+use crate::discovery::discovery::Monitor;
+
+#[derive(Debug)]
+pub struct MonitorPing {
+    pub user: String,
+    pub password: String,
+    pub ping_interval: u64,
+    pub ping_timeout: u64,
+    pub ping_max_failures: u64,
+    pub ping_tx: crossbeam_channel::Sender<PingMonitorResponse>,
+    pub rw_endpoint: ReadWriteEndpoint,
+}
+
+#[derive(Debug, Clone)]
+pub enum PingStatus {
+    PingOk,
+    PingNotOk,
+}
+
+#[derive(Debug, Clone)]
+pub struct PingMonitorResponse {
+    pub read: HashMap<String, PingStatus>,
+    pub readwrite: HashMap<String, PingStatus>,
+}
+
+impl PingMonitorResponse {
+    pub fn new(rw_endpoint: ReadWriteEndpoint) -> Self {
+        let mut read = HashMap::new();
+        let mut readwrite = HashMap::new();
+        for r in rw_endpoint.read {
+            read.insert(r.addr, PingStatus::PingNotOk);
+        }
+        for rw in rw_endpoint.readwrite {
+            readwrite.insert(rw.addr, PingStatus::PingNotOk);
+        }
+        PingMonitorResponse { read, readwrite }
+    }
+}
+
+impl MonitorPing {
+    pub fn new(
+        user: String,
+        password: String,
+        ping_interval: u64,
+        ping_timeout: u64,
+        ping_max_failures: u64,
+        ping_tx: crossbeam_channel::Sender<PingMonitorResponse>,
+        rw_endpoint: ReadWriteEndpoint,
+    ) -> Self {
+        MonitorPing {
+            user,
+            password,
+            ping_interval,
+            ping_timeout,
+            ping_max_failures,
+            ping_tx,
+            rw_endpoint,
+        }
+    }
+
+    async fn ping_check(
+        user: String,
+        password: String,
+        addr: String,
+    ) -> Result<PingStatus, Error> {
+       
+        let factory = ClientConn::with_opts(user, password, addr.clone());
+        let mut client_conn = match factory.connect().await {
+            Ok(client_conn) => client_conn,
+            Err(_) => return Ok(PingStatus::PingNotOk),
+        };
+
+        match client_conn.send_ping().await {
+            Ok(ping_ok) => {
+                if ping_ok {
+                    return Ok(PingStatus::PingOk);
+                } else {
+                    return Ok(PingStatus::PingNotOk);
+                }
+            }
+            Err(_) => {
+                return Ok(PingStatus::PingNotOk)},
+        }
+       
+    }
+}
+
+#[async_trait::async_trait]
+impl Monitor for MonitorPing {
+    async fn run_check(&self) {
+        let user = self.user.clone();
+        let password = self.password.clone();
+        let ping_interval = self.ping_interval;
+        let ping_timeout = self.ping_timeout;
+        let ping_max_failures = self.ping_max_failures;
+        let rw_endpoint = self.rw_endpoint.clone();
+        let ping_tx = self.ping_tx.clone();
+
+        let mut response = PingMonitorResponse::new(rw_endpoint.clone());
+
+        tokio::spawn(async move {
+            let mut retries = 1;
+            loop {
+            
+                if let Err(_) = time::timeout(Duration::from_millis(ping_timeout), async {
+                    for read in rw_endpoint.clone().read {
+                        match MonitorPing::ping_check(
+                            user.clone(),
+                            password.clone(),
+                            read.addr.clone(),
+                        )
+                        .await
+                        {
+                            Ok(ping_status) => match ping_status {
+                                PingStatus::PingOk => {
+                                    response.read.insert(read.addr, PingStatus::PingOk);
+                                }
+                                PingStatus::PingNotOk => loop {
+                                    if retries > ping_max_failures {
+                                        response
+                                            .read
+                                            .insert(read.addr.clone(), PingStatus::PingNotOk);
+                                        retries = 1;
+                                        break;
+                                    } else {
+                                        match MonitorPing::ping_check(
+                                            user.clone(),
+                                            password.clone(),
+                                            read.addr.clone(),
+                                        )
+                                        .await
+                                        {
+                                            Ok(ping_status) => match ping_status {
+                                                PingStatus::PingOk => {
+                                                    response.read.insert(
+                                                        read.addr.clone(),
+                                                        PingStatus::PingOk,
+                                                    );
+                                                    retries = 1;
+                                                    break;
+                                                }
+                                                PingStatus::PingNotOk => {
+                                                    retries += 1;
+                                                }
+                                            },
+                                            Err(_) => { 
+                                                retries += 1
+                                            },
+                                        }
+                                    }
+                                    std::thread::sleep(std::time::Duration::from_millis(
+                                        ping_interval,
+                                    ));
+                                },
+                            },
+                            Err(_) => {
+                                loop {
+                                    if retries > ping_max_failures {
+                                        response
+                                            .read
+                                            .insert(read.addr.clone(), PingStatus::PingNotOk);
+                                        retries = 1;
+                                        break;
+                                    } else {
+                                        match MonitorPing::ping_check(
+                                            user.clone(),
+                                            password.clone(),
+                                            read.addr.clone(),
+                                        )
+                                        .await
+                                        {
+                                            Ok(ping_status) => match ping_status {
+                                                PingStatus::PingOk => {
+                                                    response.read.insert(
+                                                        read.addr.clone(),
+                                                        PingStatus::PingOk,
+                                                    );
+                                                    break;
+                                                }
+                                                PingStatus::PingNotOk => {
+                                                    retries += 1;
+                                                }
+                                            },
+                                            Err(_) => {
+                                                retries += 1
+                                            },
+                                        }
+                                    }
+                                    std::thread::sleep(std::time::Duration::from_millis(
+                                        ping_interval,
+                                    ));
+                                }
+                            }
+                        }
+                    }
+
+                    for readwrite in rw_endpoint.clone().readwrite {
+                        match MonitorPing::ping_check(
+                            user.clone(),
+                            password.clone(),
+                            readwrite.addr.clone(),
+                        )
+                        .await
+                        {
+                            Ok(ping_status) => match ping_status {
+                                PingStatus::PingOk => {
+                                    response.readwrite.insert(readwrite.addr, PingStatus::PingOk);
+                                }
+                                PingStatus::PingNotOk => loop {
+                                    if retries > ping_max_failures {
+                                        response.readwrite.insert(
+                                            readwrite.addr.clone(),
+                                            PingStatus::PingNotOk,
+                                        );
+                                    } else {
+                                        match MonitorPing::ping_check(
+                                            user.clone(),
+                                            password.clone(),
+                                            readwrite.addr.clone(),
+                                        )
+                                        .await
+                                        {
+                                            Ok(ping_status) => match ping_status {
+                                                PingStatus::PingOk => {
+                                                    response.readwrite.insert(
+                                                        readwrite.addr.clone(),
+                                                        PingStatus::PingOk,
+                                                    );
+                                                    break;
+                                                }
+                                                PingStatus::PingNotOk => {
+                                                    retries += 1;
+                                                }
+                                            },
+                                            Err(_) => {
+                                                response.readwrite.insert(
+                                                    readwrite.addr.clone(),
+                                                    PingStatus::PingNotOk,
+                                                );
+                                            }
+                                        }
+                                    }
+                                    std::thread::sleep(std::time::Duration::from_millis(
+                                        ping_interval,
+                                    ));
+                                },
+                            },
+                            Err(_) => {
+                                continue;
+                            }
+                        }
+                    }
+                })
+                .await
+                {
+                    if retries > ping_max_failures {
+                        retries = 1;
+                    }
+                    retries += 1;
+                }
+
+                if let Err(e) = ping_tx.send(response.clone()) {
+                    // println!(">>>>>> {:#?}", e);
+                }
+                std::thread::sleep(std::time::Duration::from_millis(ping_interval));
+            }
+        });
+    }
+}
